+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 4.5.0 (Unreleased)
 
+BUG FIXES:
+
+* resource/cloudflare_ruleset: ensure custom cache keys using query parameters are defined as known values for state handling ([#2388](https://github.com/cloudflare/terraform-provider-cloudflare/issues/2388))
+
 ## 4.4.0 (19th April, 2023)
 
 NOTES: